refactor: simplify redis publisher writes
